Fix for OPENAM-626; helpers only end up in the list if they do something
git-svn-id: https://svn.forgerock.org/openam/trunk@1347 0f4defcf-c51a-4c67-9f44-6fb5eba73c5d

+++ opensso/products/amserver/source/org/forgerock/openam/upgrade/helpers/SessionServiceHelper.java
@@ -56,6 +56,8 @@ public class SessionServiceHelper extends AbstractUpgradeHelper {
             "org.forgerock.openam.session.service.DenyAccessAction";
     private final static String DESTROY_ALL_SESSIONS_CLASS = 
             "org.forgerock.openam.session.service.DestroyAllAction";
+    private final static String DESTROY_OLDEST_SESSIONS_CLASS = 
+            "org.forgerock.openam.session.service.DestroyOldestAction";
     private static final String DESTROY_ALL_SESSIONS =
             "openam.session.destroy_all_sessions";
     
@@ -74,6 +76,13 @@ public class SessionServiceHelper extends AbstractUpgradeHelper {
         }
         
         Set<String> defaultValues = existingAttr.getDefaultValues();
+        
+        if (defaultValues.contains(DESTROY_ALL_SESSIONS_CLASS) || 
+            defaultValues.contains(NEW_DENY_ACCESS) ||
+            defaultValues.contains(DESTROY_OLDEST_SESSIONS_CLASS)) {
+            // nothing to do
+            return null;
+        }
         
         if (destroyAllSessionsSet() && defaultValues.contains(DESTROY_OLD_SESSION)) {
             Set<String> newDefaultValues = new HashSet<String>();
